Move the post form all into the component in keeping with the guidelines I've been writing.
