+++ src/components/post-form/index.js
@@ -1,0 +1,16 @@
+/**
+ * External
+ */
+import Vue from 'vue';
+
+/**
+ * Internal
+ */
+import postForm from './template.vue';
+import './style.scss';
+
+/**
+ * Declare the component and export it
+ */
+Vue.component('post-form', postForm);
+export default postForm;
