+++ src/instances/post-create/post-create.vue
@@ -1,57 +1,13 @@
 <template>
-  <form
-    id="post-create-container"
-    action=""
-    method="post">
+  <section id="post-create">
     <header>
       <h1
         class="pagetitle"
         v-text="pagetitle"/>
     </header>
 
-    <div
-      id="post-title-row"
-      class="form-row">
-      <label for="post-title">Title</label>
-      <input
-        id="post-title"
-        v-model="postTitle"
-        type="text"
-        name="post-title">
-
-      <div class="generated-url-container">
-        URL: <code class="generated-url">{{ postSlug }}</code>
-      </div>
-    </div>
-
-    <div
-      id="post-excerpt-row"
-      class="form-row">
-      <label for="post-excerpt">Excerpt</label>
-      <input
-        id="post-excerpt"
-        v-model="postExcerpt"
-        type="text"
-        name="post-excerpt">
-    </div>
-
-    <div class="form-row">
-      <label for="post-body">Body</label>
-      <textarea
-        id="post-body"
-        name="post-body"
-        class="wysiwyg-editor"/>
-    </div>
-
-    <div class="form-row">
-      <button
-        class="submit-button"
-        @click="attemptStore">
-        <i class="fas fa-save"/>
-        Save
-      </button>
-    </div>
-  </form>
+    <post-form action="/admin/posts" />
+  </section>
 </template>
 
 <script>
@@ -59,41 +15,10 @@
      * Internal
      */
     import page from '../../mixins/page.js';
-
-    /**
-     * Classes
-     */
-    import Post from '../../library/post.js';
+    import postForm from '../../components/post-form';
 
     export default {
+        components: {postForm},
         mixins: [page],
-        data: () => ({
-            editor: null,
-            postBody: '',
-            postExcerpt: '',
-            postTitle: '',
-            postSlug: 'Add a title to generate the URL',
-        }),
-        mounted: function () {
-            this.editor = new SimpleMDE({
-                element: document.getElementById('post-body'),
-            });
-        },
-        methods: {
-            /**
-             * Submits data provided by the user to see if it's valid.
-             */
-            attemptStore (event) {
-                event.preventDefault();
-
-                const api = new Post();
-                api.store()
-                    .then(response => {
-                        //
-                    }).catch(response => {
-                        //
-                    });
-            },
-        },
     };
 </script>
