+++ src/instances/post-create/style.scss
@@ -1,28 +1,1 @@
 @import '../../sass/organise';
-@import '~simplemde/dist/simplemde.min.css';
-
-#post-create-container {
-    .submit-button {
-        @media screen and (min-width: 768px) {
-            width: 15%;
-            min-width: 150px;
-        }
-    }
-}
-
-#post-excerpt-row,
-#post-title-row {
-    @media screen and (min-width: 768px) {
-        width: 65%;
-    }
-}
-
-.generated-url-container {
-    padding: .5em 0 0;
-
-    .generated-url {
-        margin-left: 1em;
-        padding: 0 .5em;
-        background: $colour-background-light;
-    }
-}
